+++ src/CDend.h
@@ -2,6 +2,7 @@
 #include <Rdefines.h>
 #include <R_ext/Rdynload.h>
 #include <R_ext/Utils.h>
+#include <R_ext/Random.h>
 #include <Rinternals.h>
 #include <math.h>
 #include <stdlib.h>
@@ -38,6 +39,7 @@ void internalPartitionMap(treeNode *node, bool **pSets, unsigned int *hvs, int l
 int reallocPartitionMap(bool **pSets, int lh, int plen);
 double scorePMs(bool **pm1, bool **pm2, int pm1l, int pm2l, int lh);
 double calcEntropy(bool **pm, int lh, int pml);
+unsigned long RFHashMap(treeNode *node, unsigned long *htable, unsigned long *keys, unsigned int *hvs, int lh, int rootv);
 
 /* Fitch Parsimony */
 void resetTree(treeNode* node, int val);